Move jsonschema validations to type-specific packages
