--- src/commonMain/kotlin/io/konform/validation/string/ValidationBuilderString.kt
@@ -1,6 +1,0 @@
-package io.konform.validation.string
-
-import io.konform.validation.Constraint
-import io.konform.validation.ValidationBuilder
-
-public fun ValidationBuilder<String>.notBlank(): Constraint<String> = addConstraint("must not be blank") { it.isNotBlank() }
